--- lib/features/read_markdown_file/domain/use_cases/add_question_answer_pairs_in_note_to_ankidroid_and_get_ids_use_case.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/services.dart';
-import 'package:markdown_to_flashcard/features/read_markdown_file/domain/entities/question_answer_pair.dart';
-
-import '../entities/note.dart';
-
-class AddQuestionAnswerPairsInNoteToAnkidroidAndGetIDsUseCase {
-  final MethodChannel methodChannel;
-
-  AddQuestionAnswerPairsInNoteToAnkidroidAndGetIDsUseCase(
-      {required this.methodChannel});
-
-  Future<List<int>> call(Note note) async {
-    List<int> createdNoteIds = [];
-
-    for (QuestionAnswerPair qaPair in note.questionAnswerPairs) {
-      int createdNoteId = await methodChannel.invokeMethod(
-        'addAnkiFlashcard',
-        <String, dynamic>{
-          'deck': note.deck,
-          'question': qaPair.question,
-          'answer': qaPair.answer,
-          'source': note.title,
-          'tags': note.tags,
-        },
-      );
-
-      createdNoteIds.add(createdNoteId);
-    }
-
-    return createdNoteIds;
-  }
-}
