+++ lib/features/read_markdown_file/presentation/bloc/markdown_to_flashcard_cubit.dart
@@ -1,16 +1,16 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:markdown_to_flashcard/features/read_markdown_file/domain/use_cases/add_flashcard_ids_to_note_use_case.dart';
-import 'package:markdown_to_flashcard/features/read_markdown_file/domain/use_cases/add_question_answer_pairs_in_note_to_ankidroid_and_get_ids_use_case.dart';
 
 import '../../data/repositories/note_repository.dart';
 import '../../domain/entities/note.dart';
 import '../../domain/use_cases/convert_markdown_to_html_use_case.dart';
+import '../../domain/use_cases/create_or_update_flashcards_in_note_to_ankidroid_use_case.dart';
 import 'markdown_to_flashcard_state.dart';
 
 class MarkdownToFlashcardCubit extends Cubit<MarkdownToFlashcardState> {
   final NoteRepository noteRepository;
   final ConvertMarkdownToHTMLUseCase convertMarkdownToHTML;
-  final AddQuestionAnswerPairsInNoteToAnkidroidAndGetIDsUseCase
+  final CreateOrUpdateFlashcardsInNoteToAnkidroidUseCase
       addQuestionAnswerPairsInNoteToAnkidroidAndGetIDs;
   final AddFlashcardIDsToNoteUseCase addFlashcardIDsToNote;
